requirements.txt
streamlit
google-generativeai
requests
python-dotenv
streamlit-chat